README.md
Recru  is a web application designed to help recruiters efficiently evaluate resumes against a job description (JD) by calculating similarity scores. Recruiters can upload a JD and up to 10 resumes. The app uses Natural Language Processing (NLP) to compare each resume with the JD and rank them based on similarity. The results are displayed in an easy-to-read format with bar charts to help visualize the comparison.